devtools/5gshell/fiveshell/requirements.txt
PyYAML>=5.4.1
kubernetes>=12.0.1
requests>=2.25.1
PySide2>=5.15.2
jmespath>=0.10.0
packaging>=20.9

# Additional Packages:
# Ubuntu:
# sudo apt install python3-pip python3-venv sshpass libxml2-utils

